requirements.txt
asgiref==3.8.1
coverage==7.6.3
dj-database-url==2.2.0
Django==5.1.2
django-filter==24.3
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.7.0
